pyproject.toml
[project]
name = 'umap_dsfr'
version = '0.0.1'
dependencies = [
    'umap-project',
    'mistune',
    'python-frontmatter',
    'python-jose',
]

[project.optional-dependencies]
dev = [
    'black==21.10b0',
    'djlint==1.31.0',
]

[tool.setuptools]
packages = [
    'umap_dsfr',
]

[tool.djlint]
profile="django"
indent=2
format_css=true
